added default image v:shape attribute "stroked" = "false". remove image strokes on other editors such as onlyoffice
the option of able to set the stroke of the image is not available in the PHPWord. But even not defining the stroke, other Editors will set a default Stroke in black such as in Onlyoffice. It can only be deactivated if an attribute "stroked" = "false" on the v:shape is inserted.